requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
numpy
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
twilio
